robot/requirements.txt
pyyaml==6.0
robotframework==6.0
robotframework-selenium2library==3.0.0
robotframework-httplibrary==0.4.2
robotframework-jsonlibrary==0.5
robotframework-lint==1.1
robotframework-requests==0.9.3
requests==2.31.0
webdrivermanager==0.10.0
RESTinstance==1.3.0
jsonschema==4.17.0
robotframework-excellentlibrary==1.0.0
robotframework-faker==5.0.0
Faker==15.3.4


